backend/requirements.txt
Django==5.0.2
djangorestframework==3.15.1
django-cors-headers==4.3.1
django-filter==24.2

# Utility & time handling
python-dateutil==2.9.0
pytz==2024.1

# For running production servers
gunicorn==21.2.0

# Optional: Testing & Coverage
pytest==8.2.2
pytest-django==4.8.0
coverage==7.4.1
